Broke out models into separate classes
