+++ AdventureGame/Models/Character.cs
@@ -1,0 +1,9 @@
+﻿namespace AdventureGame.Models
+{
+    public abstract class Character
+    {
+        public string Name { get; set; }
+        public string Type { get; set; }
+        public int Health { get; set; }
+    }
+}
