+++ AdventureGame/Models/Hero.cs
@@ -1,0 +1,7 @@
+﻿namespace AdventureGame.Models
+{
+    public class Hero : Character
+    {
+
+    }
+}
