+++ AdventureGame/Models/Item.cs
@@ -1,0 +1,7 @@
+﻿namespace AdventureGame.Models
+{
+    public abstract class Item
+    {
+        public string Name { get; set; }
+    }
+}
