+++ AdventureGame/Models/Weapon.cs
@@ -1,0 +1,8 @@
+﻿namespace AdventureGame.Models
+{
+    public class Weapon : Item
+    {
+        public int Damage { get; set; }
+        public int Durability { get; set; }
+    }
+}
