+++ AdventureGame/Program.cs
@@ -1,5 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
+using System;
+using AdventureGame.Models;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -31,87 +31,10 @@ namespace AdventureGame
                         break;
                     default:
                         Console.WriteLine("Thank you for playing the game");
-                        Game.Playing = false;
+                        game.Playing = false;
                         break;
                 }
             }
-        }
-    }
-
-    class Game {
-        public bool Playing
-        {
-            get; set;
-        }
-        public void GameOver() {
-            
-        }
-        public void Save() { }
-        public void goToTown( string TownName )
-        {
-            Console.WriteLine("Welcome to " + TownName + ". What would you like to do? \n"
-                               + "\t 1.) Shop\n"
-                               + "\t 2.) Heal\n"
-                               + "\t 3.) Rest\n" 
-                               + "\t 4.) Save");
-            switch (Console.ReadLine())
-            {
-                case "1":
-                    goToShop();
-                    break;
-                case "2":
-                    goToHeal();
-                    break;
-                case "3":
-                    goToRest();
-                    break;
-                case "4":
-                    Save();
-                    break;
-                default:
-                    break;
-            }
-        }
-        private void goToShop()
-        { }
-        private void goToHeal()
-        { }
-        private void goToRest()
-        { }
-    }
-    class Character {
-        public string Name
-        {
-            get; set;
-        }
-        public string Type
-        {
-            get; set;
-        }
-        public int Health
-        {
-            get; set;
-        }
-    }
-    class Hero : Character {
-
-    }
-    class Enemy : Character {
-
-    }
-    class Item {
-        public string Name
-        {
-            get; set;
-        }
-    }
-    class Weapon : Item {
-        public int Damage {
-            get; set;
-        }
-        public int Durability
-        {
-            get; set;
         }
     }
 }
